ファイル分割してreactの役割ごとに整理した。
