+++ my-app2/src/actions/tasks.js
@@ -1,0 +1,15 @@
+// 初回のタスクを追加するためのオブジェクトかな。
+export const inputTask = (task) => ({
+  type: 'INPUT_TASK',
+  payload: {
+    task
+  },
+});
+
+// タスクを追加するためのオブジェクトかな。
+export const addTask = (task) => ({
+  type: 'ADD_TASK',
+  payload: {
+    task
+  },
+});
